+++ src/ReadonlyMap.ts
@@ -71,18 +71,14 @@ export function getShow<K, A>(SK: Show<K>, SA: Show<A>): Show<ReadonlyMap<K, A>>
  *
  * @since 2.5.0
  */
-export function size<K, A>(d: ReadonlyMap<K, A>): number {
-  return d.size
-}
+export const size = (m: ReadonlyMap<unknown, unknown>): number => m.size
 
 /**
  * Test whether or not a map is empty
  *
  * @since 2.5.0
  */
-export function isEmpty<K, A>(d: ReadonlyMap<K, A>): boolean {
-  return d.size === 0
-}
+export const isEmpty = (m: ReadonlyMap<unknown, unknown>): boolean => m.size === 0
 
 // TODO: remove non-curried overloading in v3
 /**